3.1.0

+++ LiveContainerSwiftUI/Shared.swift
@@ -9,6 +9,7 @@ import SwiftUI
 import UniformTypeIdentifiers
 import LocalAuthentication
 import SafariServices
+import Security
 
 struct LCPath {
     public static let docPath = {
@@ -53,8 +54,15 @@ struct LCPath {
 
 class SharedModel: ObservableObject {
     @Published var isHiddenAppUnlocked = false
+    @Published var developerMode = false
     // 0= not installed, 1= is installed, 2=current liveContainer is the second one
     @Published var multiLCStatus = 0
+    
+    @Published var apps : [LCAppModel] = []
+    @Published var hiddenApps : [LCAppModel] = []
+    let isPhone: Bool = {
+        UIDevice.current.userInterfaceIdiom == .phone
+    }()
     
     func updateMultiLCStatus() {
         if LCUtils.appUrlScheme()?.lowercased() != "livecontainer" {
@@ -116,12 +124,12 @@ class InputHelper : AlertHelper<String> {
 extension String: LocalizedError {
     public var errorDescription: String? { return self }
         
-    private static var enBundle : Bundle? {
+    private static var enBundle : Bundle? = {
         let language = "en"
         let path = Bundle.main.path(forResource:language, ofType: "lproj")
         let bundle = Bundle(path: path!)
         return bundle
-    }
+    }()
     
     var loc: String {
         let message = NSLocalizedString(self, comment: "")
@@ -175,6 +183,16 @@ extension View {
         self.modifier(TextFieldAlertModifier(isPresented: isPresented, title: title, text: text, placeholder: placeholder, action: action, actionCancel: actionCancel))
     }
     
+    public func betterFileImporter(
+        isPresented: Binding<Bool>,
+        types : [UTType],
+        multiple : Bool = false,
+        callback: @escaping ([URL]) -> (),
+        onDismiss: @escaping () -> Void
+    ) -> some View {
+        self.modifier(DocModifier(isPresented: isPresented, types: types, multiple: multiple, callback: callback, onDismiss: onDismiss))
+    }
+    
     func onBackground(_ f: @escaping () -> Void) -> some View {
         self.onReceive(
             NotificationCenter.default.publisher(for: UIApplication.willResignActiveNotification),
@@ -189,6 +207,70 @@ extension View {
         )
     }
     
+}
+
+public struct DocModifier: ViewModifier {
+
+    @State private var docController: UIDocumentPickerViewController?
+    @State private var delegate : UIDocumentPickerDelegate
+    
+    @Binding var isPresented: Bool
+
+    var callback: ([URL]) -> ()
+    private let onDismiss: () -> Void
+    private let types : [UTType]
+    private let multiple : Bool
+    
+    init(isPresented : Binding<Bool>, types : [UTType], multiple : Bool, callback: @escaping ([URL]) -> (), onDismiss: @escaping () -> Void) {
+        self.callback = callback
+        self.onDismiss = onDismiss
+        self.types = types
+        self.multiple = multiple
+        self.delegate = Coordinator(callback: callback, onDismiss: onDismiss)
+        self._isPresented = isPresented
+    }
+
+    public func body(content: Content) -> some View {
+        content.onChange(of: isPresented) { isPresented in
+            if isPresented, docController == nil {
+                let controller = UIDocumentPickerViewController(forOpeningContentTypes: types, asCopy: true)
+                controller.allowsMultipleSelection = multiple
+                controller.delegate = delegate
+                self.docController = controller
+                guard let scene = UIApplication.shared.connectedScenes.first as? UIWindowScene else {
+                    return
+                }
+                scene.windows.first?.rootViewController?.present(controller, animated: true)
+            } else if !isPresented, let docController = docController {
+                docController.dismiss(animated: true)
+                self.docController = nil
+            }
+        }
+    }
+
+    private func shutdown() {
+        isPresented = false
+        docController = nil
+    }
+    
+    class Coordinator: NSObject, UIDocumentPickerDelegate {
+        var callback: ([URL]) -> ()
+        private let onDismiss: () -> Void
+        
+        init(callback: @escaping ([URL]) -> Void, onDismiss: @escaping () -> Void) {
+            self.callback = callback
+            self.onDismiss = onDismiss
+        }
+        
+        func documentPicker(_ controller: UIDocumentPickerViewController, didPickDocumentsAt urls: [URL]) {
+            callback(urls)
+            onDismiss()
+        }
+        func documentPickerWasCancelled(_ controller: UIDocumentPickerViewController) {
+            onDismiss()
+        }
+    }
+
 }
 
 public struct TextFieldAlertModifier: ViewModifier {
@@ -306,9 +388,9 @@ struct SiteAssociation : Codable {
 }
 
 extension LCUtils {
-    public static let appGroupUserDefault = UserDefaults.init(suiteName: LCUtils.appGroupID())!
+    public static let appGroupUserDefault = UserDefaults.init(suiteName: LCUtils.appGroupID()) ?? UserDefaults.standard
     
-    public static func signFilesInFolder(url: URL, onProgressCreated: (Progress) -> Void) async -> String? {
+    public static func signFilesInFolder(url: URL, signer:Signer, onProgressCreated: (Progress) -> Void) async -> (String?, Date?) {
         let fm = FileManager()
         var ans : String? = nil
         let codesignPath = url.appendingPathComponent("_CodeSignature")
@@ -322,23 +404,35 @@ extension LCUtils {
         do {
             try fm.copyItem(at: Bundle.main.executableURL!, to: tmpExecPath)
         } catch {
-            return nil
+            return (nil, nil)
         }
-        
+        var ansDate : Date? = nil
         await withCheckedContinuation { c in
-            let progress = LCUtils.signAppBundle(url) { success, error in
+            func compeletionHandler(success: Bool, expirationDate: Date?, error: Error?){
                 do {
                     if let error = error {
                         ans = error.localizedDescription
                     }
-                    try fm.removeItem(at: codesignPath)
-                    try fm.removeItem(at: provisionPath)
+                    if(fm.fileExists(atPath: codesignPath.path)) {
+                        try fm.removeItem(at: codesignPath)
+                    }
+                    if(fm.fileExists(atPath: provisionPath.path)) {
+                        try fm.removeItem(at: provisionPath)
+                    }
+
                     try fm.removeItem(at: tmpExecPath)
                     try fm.removeItem(at: tmpInfoPath)
+                    ansDate = expirationDate
                 } catch {
                     ans = error.localizedDescription
                 }
                 c.resume()
+            }
+            let progress : Progress?
+            if signer == .AltSign {
+                progress = LCUtils.signAppBundle(url, completionHandler: compeletionHandler)
+            } else {
+                progress = LCUtils.signAppBundle(withZSign: url, completionHandler: compeletionHandler)
             }
             guard let progress = progress else {
                 ans = "lc.utils.initSigningError".loc
@@ -347,8 +441,131 @@ extension LCUtils {
             }
             onProgressCreated(progress)
         }
-        return ans
+        return (ans, ansDate)
 
+    }
+    
+    public static func signTweaks(tweakFolderUrl: URL, force : Bool = false, signer:Signer, progressHandler : ((Progress) -> Void)? = nil) async throws {
+        guard LCUtils.certificatePassword() != nil else {
+            return
+        }
+        let fm = FileManager.default
+        var isFolder :ObjCBool = false
+        if(fm.fileExists(atPath: tweakFolderUrl.path, isDirectory: &isFolder) && !isFolder.boolValue) {
+            return
+        }
+        
+        // check if re-sign is needed
+        // if sign is expired, or inode number of any file changes, we need to re-sign
+        let tweakSignInfo = NSMutableDictionary(contentsOf: tweakFolderUrl.appendingPathComponent("TweakInfo.plist")) ?? NSMutableDictionary()
+        let expirationDate = tweakSignInfo["expirationDate"] as? Date
+        var signNeeded = false
+        if let expirationDate, expirationDate.compare(Date.now) == .orderedDescending, !force {
+            
+            let tweakFileINodeRecord = tweakSignInfo["files"] as? [String:NSNumber] ?? [String:NSNumber]()
+            let fileURLs = try fm.contentsOfDirectory(at: tweakFolderUrl, includingPropertiesForKeys: nil)
+            for fileURL in fileURLs {
+                let attributes = try fm.attributesOfItem(atPath: fileURL.path)
+                let fileType = attributes[.type] as? FileAttributeType
+                if(fileType != FileAttributeType.typeDirectory && fileType != FileAttributeType.typeRegular) {
+                    continue
+                }
+                if(fileType == FileAttributeType.typeDirectory && !fileURL.lastPathComponent.hasSuffix(".framework")) {
+                    continue
+                }
+                if(fileType == FileAttributeType.typeRegular && !fileURL.lastPathComponent.hasSuffix(".dylib")) {
+                    continue
+                }
+                
+                if(fileURL.lastPathComponent == "TweakInfo.plist"){
+                    continue
+                }
+                let inodeNumber = try fm.attributesOfItem(atPath: fileURL.path)[.systemFileNumber] as? NSNumber
+                if let fileInodeNumber = tweakFileINodeRecord[fileURL.lastPathComponent] {
+                    if(fileInodeNumber != inodeNumber) {
+                        signNeeded = true
+                        break
+                    }
+                } else {
+                    signNeeded = true
+                    break
+                }
+                
+                print(fileURL.lastPathComponent) // Prints the file name
+            }
+            
+        } else {
+            signNeeded = true
+        }
+        
+        guard signNeeded else {
+            return
+        }
+        // sign start
+        
+        let tweakItems : [String] = []
+        let tmpDir = fm.temporaryDirectory.appendingPathComponent("TweakTmp.app")
+        if fm.fileExists(atPath: tmpDir.path) {
+            try fm.removeItem(at: tmpDir)
+        }
+        try fm.createDirectory(at: tmpDir, withIntermediateDirectories: true)
+        
+        var tmpPaths : [URL] = []
+        // copy items to tmp folders
+        let fileURLs = try fm.contentsOfDirectory(at: tweakFolderUrl, includingPropertiesForKeys: nil)
+        for fileURL in fileURLs {
+            let attributes = try fm.attributesOfItem(atPath: fileURL.path)
+            let fileType = attributes[.type] as? FileAttributeType
+            if(fileType != FileAttributeType.typeDirectory && fileType != FileAttributeType.typeRegular) {
+                continue
+            }
+            if(fileType == FileAttributeType.typeDirectory && !fileURL.lastPathComponent.hasSuffix(".framework")) {
+                continue
+            }
+            if(fileType == FileAttributeType.typeRegular && !fileURL.lastPathComponent.hasSuffix(".dylib")) {
+                continue
+            }
+            
+            let tmpPath = tmpDir.appendingPathComponent(fileURL.lastPathComponent)
+            tmpPaths.append(tmpPath)
+            try fm.copyItem(at: fileURL, to: tmpPath)
+        }
+        
+        if tmpPaths.isEmpty {
+            try fm.removeItem(at: tmpDir)
+            return
+        }
+        
+        let (error, expirationDate2) = await LCUtils.signFilesInFolder(url: tmpDir, signer: signer) { p in
+            if let progressHandler {
+                progressHandler(p)
+            }
+        }
+        if let error = error {
+            throw error
+        }
+        
+        // move signed files back and rebuild TweakInfo.plist
+        tweakSignInfo.removeAllObjects()
+        tweakSignInfo["expirationDate"] = expirationDate2
+        var fileInodes = [String:NSNumber]()
+        for tmpFile in tmpPaths {
+            let toPath = tweakFolderUrl.appendingPathComponent(tmpFile.lastPathComponent)
+            // remove original item and move the signed ones back
+            if fm.fileExists(atPath: toPath.path) {
+                try fm.removeItem(at: toPath)
+                
+            }
+            try fm.moveItem(at: tmpFile, to: toPath)
+            if let inodeNumber = try fm.attributesOfItem(atPath: toPath.path)[.systemFileNumber] as? NSNumber {
+                fileInodes[tmpFile.lastPathComponent] = inodeNumber
+            }
+        }
+        try fm.removeItem(at: tmpDir)
+
+        tweakSignInfo["files"] = fileInodes
+        try tweakSignInfo.write(to: tweakFolderUrl.appendingPathComponent("TweakInfo.plist"))
+        
     }
     
     public static func getAppRunningLCScheme(bundleId: String) -> String? {
@@ -361,6 +578,26 @@ extension LCUtils {
         // Iterate over the dictionary to find the matching bundle ID
         for (key, value) in info {
             if value == bundleId {
+                if key == LCUtils.appUrlScheme() {
+                    return nil
+                }
+                return key
+            }
+        }
+        
+        return nil
+    }
+    
+    public static func getContainerUsingLCScheme(containerName: String) -> String? {
+        // Retrieve the app group path using the app group ID
+        let infoPath = LCPath.lcGroupDocPath.appendingPathComponent("containerLock.plist")
+        // Read the plist file into a dictionary
+        guard let info = NSDictionary(contentsOf: infoPath) as? [String: String] else {
+            return nil
+        }
+        // Iterate over the dictionary to find the matching bundle ID
+        for (key, value) in info {
+            if value == containerName {
                 if key == LCUtils.appUrlScheme() {
                     return nil
                 }
@@ -442,4 +679,18 @@ extension LCUtils {
             return "Unknown Store"
         }
     }
+    
+    public static func removeAppKeychain(dataUUID label: String) {
+        [kSecClassGenericPassword, kSecClassInternetPassword, kSecClassCertificate, kSecClassKey, kSecClassIdentity].forEach {
+          let status = SecItemDelete([
+            kSecClass as String: $0,
+            "alis": label,
+          ] as CFDictionary)
+          if status != errSecSuccess && status != errSecItemNotFound {
+              //Error while removing class $0
+              NSLog("[LC] Failed to find keychain items: \(status)")
+          }
+        }
+    }
+
 }
